Fix: graph progress is incorrect when graph failed, timeout or cancelled.
The original issue is that  graph progress is incorrect when graph
failed, timeout or cancelled.

The root cause is that the method `tap` in Rxjs is not waited to finish,
and the data flowing into tap is modified by later stream operation.

The solution is `cloneDeep` the `object` to be published.

Recently, https://github.com/RackHD/on-taskgraph/pull/278 has
`cloneDeeped` the `graph`, so the problem has been solved in fact. So
this updated PR is to `cloneDeep` the remain `objects` to be published.